session issue, properties done<?>
