+++ src/main/java/pl/edu/ur/pnes/editor/Session.java
@@ -6,6 +6,7 @@ import javafx.scene.control.Tab;
 import pl.edu.ur.pnes.editor.history.UndoHistory;
 import pl.edu.ur.pnes.petriNet.Net;
 import pl.edu.ur.pnes.petriNet.PetriNet;
+import pl.edu.ur.pnes.ui.panels.CenterPanelController;
 
 import java.io.File;
 
@@ -33,6 +34,7 @@ public class Session {
     public File getFile() {
         return file.get();
     }
+
     public void setFile(File file) {
         this.file.set(file);
     }
@@ -51,6 +53,7 @@ public class Session {
     public boolean isModified() {
         return modified.get();
     }
+
     public void setModified(boolean modified) {
         this.modified.set(modified);
     }
@@ -129,4 +132,19 @@ public class Session {
     }
 
     // TODO: load (from static function/factory)
+
+
+    ////////////////////////////////////////////////////////////////////////////////
+    private CenterPanelController centerPanelController;
+
+    public CenterPanelController getCenterPanelController() {
+        return centerPanelController;
+    }
+
+    public void setCenterPanelController(CenterPanelController centerPanelController) {
+
+        this.centerPanelController = centerPanelController;
+        System.out.println(centerPanelController);
+    }
+    ////////////////////////////////////////////////////////////////////////////////
 }
